nfpm.native_libs: prevent inferred deps on script deps
The rules should not depend on `elfdeps` (a pypi package), because that
will make the pants wheel/pex include `elfdeps` which is not useful.
So, instead of depending on the script `python_source` targets, add
deps on `resource` targets instead.

Each script is both `python_source` and `resource` so that rule code can
depend on the script as a resource, and ruff, mypy, pytest, and other
tools can work with the script as python. To avoid inadvertent
dependencies between rule python code and script python code, this adds
some `__dependents_rules__` to ensure rule code can only depend on the
`resource`, not the `python_source`.