plumb through (most) of Pex's --scie flags (#22866)
After playing with some back and forth I decided to add these to the
existing `pex_binary` target: that is most similar to what Pex does; and
it ended up being the easiest way to re-use existing machinery. I tried
to stay out of Pex's way with the various flags and pass things through.
The help text is lightly translated (ex: `-flags` too fields). The most
fiddly part is the expected outputs where the existing Pants code really
wants to know the expected outputs of a Process. I effectively
duplicated some of the Pex output logic, this is annoying, but I think
mostly getting in our own way instead of limiting what anyone can do,
and it does provide nicer output.

Flags not currently covered as fields:
* `--scie-only, --no-scie-only, --pex-and-scie`: The existing Pant
classes really expect to get a `.pex` out of all of this.
* `--scie-pypy-release`: I don't have any PyPy experience. I think it
would be straightforward to add.
* `--scie-pbs-free-threaded` / `--scie-pbs-debug`: This are new since I
started; thanks! Happy to do as a followup PR.
* `--scie-science-binary`: Would maybe be interesting if someone wrote a
general `science` backend.
 * `--scie-assets-base-url` / `--scie-base`

ref #22654