nfpm.native_libs: new backend for nfpm pkg deps (only elfdeps subsystem) (#22873)
## PR Series Overview

This is the first in a series of PRs that introduces a new backend:
`pants.backend.npm.native_libs`
Initially, the backend will be available as:
`pants.backend.experimental.nfpm.native_libs`

I proposed this new backend (originally named `bindeps`) in discussion
#22396.

This backend will inspect ELF bin/lib files (like `lib*.so`) in packaged
contents (for this PR series, only in `pex_binary` targets) to identify
package dependency metadata and inject that metadata on the relevant
`nfpm_deb_package` or `nfpm_rpm_package` targets. Effectively, it will
provide an approximation of these native packager features:
- `rpm`: `rpmdeps` + `elfdeps`
- `deb`: `dh_shlibdeps` + `dpkg-shlibdeps` (These substitute
`${shlibs:Depends}` in debian control files have)

### Goal: Host-agnostic package builds

This pants backend is designed to be host-agnostic, like
[nFPM](https://nfpm.goreleaser.com/).

Native packaging tools are often restricted to a single release of a
single distro. Unlike native package builders, this new pants backend
does not use any of those distro-specific or distro-release-specific
utilities or local package databases. This new backend should be able to
build deb and rpm packages anywhere that pants can run (MacOS, rpm linux
distros, deb linux distros, other linux distros, docker, ...).

## PR Overview

To achieve the host-agnostic goal, the scripts in this new backend use
pure-python deps to search the ELF bin/lib files for provided and/or
required SONAMEs:

- `elfdeps` ([:package:](https://pypi.org/project/elfdeps/) |
[:octocat:](https://github.com/python-wheel-build/elfdeps)):
> "Python implementation of RPM
[elfdeps](https://github.com/rpm-software-management/rpm/blob/master/tools/elfdeps.c)."

- `pyelftools` ([:package:](https://pypi.org/project/pyelftools/) |
[:octocat:](https://github.com/eliben/pyelftools)):
> "A pure-Python library for parsing and analyzing ELF files and DWARF
debugging information."
(`elfdeps` is only host-agnostic because of its dependency on
`pyelftools`. We do not use `pyelftools` directly.)

This PR focuses only on `pants.backend.npm.native_libs.elfdeps`, which
includes some rules, a subsystem, a lockfile, and an `analyze.py`
script. Future PRs will actually translate and inject these SONAME
provides/requires into the appropriate `nfpm_*_package` dependency
fields.

### Subsystem and Lockfile

`elfdeps` is a subsystem (with a default `elfdeps.lock` file) that
allows configuring an alternate version of `elfdeps`. This might be
useful if an alternate version (or a fork?) has fixes/features that
apply to libraries contained in the wheels. As with all other backends,
we only test with one version in CI, so users may encounter issues with
other versions, which they are welcome to report on GitHub or in Slack.

The subsystem's new lockfile summary (note the minimal number of deps,
all of which are pure-python):
```
__________________________________________________________________
Lockfile diff: elfdeps.lock [elfdeps]
__________________________________________________________________
==                      Added dependencies                      ==
‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾
  elfdeps                        0.2.0
  pyelftools                     0.32
```

There is also a rule bundled with the subsystem,
`setup_elfdeps_analyze_wheels_tool`, that constructs a venv for the
`analyze.py` script to run in.

### Rules

The `elfdeps_analyze_pex` rule:
- Run `pex3 repository extract` to create a directory with all of the
wheels in the pex.
- Unzip the pex to get any remaining files that are not included in
wheels.
_aside: I would rather not unpack the pex like this, as the pex-internal
structure feels like an internal implementation detail. However, I
couldn't find any official pex utilities to extract the non-wheel
(non-deps) files, so `UnzipBinary` it is._
- `subsystem.setup_elfdeps_analyze_wheels_tool`: Prepare a pex venv
using `elfdeps.lock`
- Run `analyze.py` script (see below) in a pex venv. This runs
(concurrently) twice: once with the wheels, and once with the non-wheel
files.
- Parse the JSON output of `analyze.py`, returning it in a dataclass for
use by the calling rule.

### `analyze.py` script

The `analyze.py` script in `pants.backend.nfpm.native_libs.elfdeps` does
the following:
- Branch based on the `--mode` arg (`wheels` and `files` are separate
modes because `elfdeps` can descend into zip files, but it can't tell
that a `.whl` is a zip file.)
- `--mode wheels`: open each wheel as a zip file, then pass it to
`elfdeps` to iterate over and inspect the wheel's contents,
- `--mode files`: pass directory to `elfdeps` to recursively inspect all
files in that directory (descending into tar or zip files if found),
- `elfdeps` analyzes ELF metadata of each file that looks like a `.so`
library or an executable ELF binary
- Collect SONAMEs in `requires` and `provides` ELF metadata returned by
elfdeps
- Return collected results as JSON (rules can use this as simple
structured data)

### Tests

Python scripts in other backends are often treated as a resource, so
linters are either skipped entirely for them, or mypy is instructed to
ignore the 3rd party imports. I want the full help of all of these tools
to keep my code clean and catch as many issues as possible. I also want
unit tests for the parts of the script, but that means pytest needs to
be able to import the script, so its dependencies have to be available
in the python resolve that includes the `analyze.py` script.

I don't want to create another resolve just for this one script and its
tests. So, I added `elfdeps` to the `python-default` resolve, and added
visibility rules that ensures the dep is only ever used by the script
and its tests, and that the script (and its `elfdeps` dep) is only a
resource dependency (not a python dependency) of rule code.

Here is the lockfile diff from regenerating the python-default resolve's
lockfile (note the package upgrades are unrelated to elfdeps):
```
__________________________________________________________________
Lockfile diff: 3rdparty/python/user_reqs.lock [python-default]
__________________________________________________________________
==                    Upgraded dependencies                     ==
‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾
  graphql-core                   3.2.6        -->   3.2.7
  pbr                            7.0.1        -->   7.0.3
  pydantic                       2.12.3       -->   2.12.4
  pydantic-core                  2.41.4       -->   2.41.5
__________________________________________________________________
==                      Added dependencies                      ==
‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾
  elfdeps                        0.2.0
  pyelftools                     0.32
```

At 18.5 kB and 188.5 kB, the pure-python wheels for `elfdeps` and
`pyelftools` are remarkably small. However, this should only matter
during development, because the visibility rules should ensure that
`elfdeps` never becomes a transitive dep of the pants wheel or pants
pex.