[pants_ng] Next-gen options reader. (#22880)
#22862 implemented NG config. 

This PR adds support for env vars and flags, and pulls them
all together into a new `OptionsReader`.

In practice this re-uses a lot of code from the current options crate:
- The new `OptionsReader` just wraps a current `OptionParser`, but
   skips the latter's complicated bootstrapping dance.
- Env support is directly taken from the existing code.
- Flag support is reimplemented, as NG flags are handled differently.

The difference in flag handling is quite important. Currently flags
have ambiguity: given `--foo-bar-baz`, is this setting a global option
`foo_bar_baz`, or an option `bar_baz` in scope `foo`, or an option 
`baz` in scope `foo-bar`? So the existing code has to iterate
over all flags and see if they match what we're looking for.

New flags use dashes in scope names, and to separate scope
names from option names, but the latter must use underscores as word
separators. So `--foo_bar_baz`, `--foo-bar_baz` and `--foo-bar_baz`
are unambiguous. Therefore we can simply look up flags by scope and
option name in nested HashMaps.

Nothing in this PR is wired up to anything yet, so this has no impact
on end users. It is speculative work to support future NG development.