Abstract PosixFS away. (#22915)
On Windows we will want some other struct to provide
filesystem operations.

We move the PosixFS code into a module, and 
conditionally expose it under the name `FS`. 

Note that hiding the functionality behind a trait would
be a much bigger change, because that trait would not
be `dyn` compatible, due to various functions returning
`impl Trait`. However this is also not necessary since 
a binary will only have one universal implementation.