nfpm.native_libs.elfdeps: update test dep
setproctitle is used to test inspecting ELF deps in wheels.
v1.3.6 had wheels for py3.11, but not py3.14.
v1.3.7 has wheels for both py3.11 and py3.14 (using same manylinux).

The version of libc linked in the library is constrained by the
manylinux tag of the wheels. So, upgrading the actual libc deps should
only have to be upgraded in this test if the manylinux tag gets updated.

I also added "--no-build" to the test pex_binary extra args to make sure
we don't get spurious errors if someone runs this test on a platform
that does not have setproctitle wheels available. This would have
changed the error received when testing v1.3.6 with py3.14, making it
complain about the lack of a wheel instead of checking the ELF deps
of a wheel built locally from an sdist.

v1.3.7 stopped building i686/i386 wheels, so I dropped that from the
test metadata as well. We don't test on i686 anyway, so no loss.