nfpm.native_libs.deb: refactor script error handling
The API returns success (200) when returning an empty result set.
A not-found result is not a failure, as it means that the dep has to be
provided by a custom package; it just means the dep is not in one of the
distro-provided packages. So, now the script also returns success (0)
when an empty result-set is returned.