flag non-runnable targets used with `code_quality_tool` (Cherry-pick of #22875) (#22877)
Raise an exception when a non-runnable target is set as the `runnable`
for `code_quality_tool` (and other similar uses). Otherise, the user
sees an obtuse `IndexError`.

Before:

```
IndexError: tuple index out of range
```

After:

```
pants.engine.target.InvalidFieldException: The `runnable` field in target `pants-macros:gofmt_tool` must be set to the address of a target which can be run by the `run` goal. Instead, the `runnable` field is set to `:gofmt` which refers to the `adhoc_tool` target at `pants-macros:gofmt` that cannot be run by the `run` goal.
```

This particular error arose while I was trying to run `gofmt` as a
`code_quality_tool`.

Co-authored-by: Tom Dyas <tom@shoalsoftware.com>