support setting the working directory on interactive process (#22802)
We've run into a few cases where setting the working directory on an
InteractiveProcess would be useful, in particular for rules are tests
(ie TestRequest). In order to support debugging we convert the setup
process to InteractiveProcess but since that drops the working directory
you're unable to utilize it currently.

As a workaround we currently write a bash script into the sandbox like 

```
cd ${DIR} && exec {COMMAND}
```

This will help clean that up

This is also a pre-requisite to supporting setting the working dir in
the Run goal (small thread here
https://pantsbuild.slack.com/archives/C01CQHVDMMW/p1707448457657149)
I'll handle that in a separate pr though